LATTICE-1666 return count for deleting linking feedback, add admin check for deletion

+++ linker/src/main/kotlin/com/openlattice/linking/controllers/LinkingFeedbackController.kt
@@ -182,9 +182,10 @@ constructor(
         )
     }
 
-    @DeleteMapping( path = [LinkingFeedbackApi.FEEDBACK] )
-    override fun deleteLinkingFeedback(@RequestBody entityPair: EntityKeyPair) {
-        feedbackService.deleteLinkingFeedback(entityPair)
+    @DeleteMapping(path = [LinkingFeedbackApi.FEEDBACK])
+    override fun deleteLinkingFeedback(@RequestBody entityPair: EntityKeyPair): Int {
+        ensureAdminAccess() // currently it's only used by tests
+        return feedbackService.deleteLinkingFeedback(entityPair)
     }
